storage: remove error on snapshotting large ranges
Now that we have streaming snapshots, this doesn't seem valuable.

Updates #9545.
